revert convert command, fetch output format & encoding
